use integer for the fader level message

+++ src/client.cpp
@@ -198,8 +198,8 @@ CClient::CClient ( const quint16  iPortNumber,
         this, SLOT ( OnSndCrdReinitRequest ( int ) ) );
 
     QObject::connect ( &Sound,
-        SIGNAL ( ControllerInFaderLevel ( int, double ) ),
-        SIGNAL ( ControllerInFaderLevel ( int, double ) ) );
+        SIGNAL ( ControllerInFaderLevel ( int, int ) ),
+        SIGNAL ( ControllerInFaderLevel ( int, int ) ) );
 
     QObject::connect ( &Socket, SIGNAL ( InvalidPacketReceived ( CHostAddress ) ),
         this, SLOT ( OnInvalidPacketReceived ( CHostAddress ) ) );
